requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
deep-translator==1.11.4
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
